pyproject.toml
[tool.poetry]
name = "eth-contracts"
version = "0.1.0"
description = ""
authors = ["James Key <james@chainflip.io>", "Tom Nash <tom@chainflip.io>"]

[tool.poetry.dependencies]
python = "^3.7.3"
umbral = "^0.1.3-alpha.2"
eth-brownie = "^1.16.2"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
